requirements.txt
# 피지컬 AI 시스템 의존성 - PHI-3.5 내장

# 딥러닝 프레임워크 (PHI-3.5 지원)
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0  # PHI-3.5 지원 버전
accelerate>=0.24.0   # 모델 로딩 최적화
numpy>=1.24.0
sentencepiece>=0.1.99  # 토큰화

# 로보틱스 및 물리 시뮬레이션
pybullet>=3.2.5
gym>=0.21.0
stable-baselines3>=2.0.0
mujoco>=2.3.0

# 강화학습 및 메타학습
ray[rllib]>=2.5.0
optuna>=3.2.0

# 컴퓨터 비전
opencv-python>=4.7.0
pillow>=9.5.0
scikit-image>=0.20.0

# 데이터 처리 및 분석
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 메모리 및 벡터 데이터베이스
redis>=4.5.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# 네트워킹 및 통신
aiohttp>=3.8.0
websockets>=11.0.0
paho-mqtt>=1.6.0

# 유틸리티
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.3.0
tqdm>=4.65.0
asyncio-mqtt>=0.13.0

# 테스트 및 개발
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# 시스템 모니터링
psutil>=5.9.0
prometheus-client>=0.16.0

# 타임시리즈 데이터
influxdb-client>=1.36.0

# 그래프 및 시각화
networkx>=3.1.0
plotly>=5.14.0
dash>=2.10.0
